Update patch set 2
Patch Set 2:

> LGTM
 > 
 > Did you actually used it? Was it useful?

I've only toyed with it so far, but it seems like it will be very useful. I was looking at it as a way to verify that the GC scheduler is having the right effect on heap growth and CPU utilization and to compare it against the effects of the current "scheduler".

I was also thinking of using it (potentially restricted to certain events) to compute precise minimum mutator utilization plots and mutator utilization topographs. Right now just the information printed by GCprinttimes is enough to do this, but once background GC is preemptible and potentially parallel and mutators are assisting the GC, I'll need a much more precise trace to build good MMUs and MUTs. (At that point we can probably also get rid of GCprinttimes and fold some of that information in with GODEBUG=gctrace=1, but that's further out.)

 > One potential problem with tracer is that it generates too large
 > traces to study them. Probably we can capture the trace for just
 > the final benchmarking round (with the final value of N). But I did
 > not use tracer on these benchmarks yet, so I don't know how useful
 > it will be.

The size seemed manageable in my admittedly small experiments. What about keeping the whole trace and using user trace events to demarcate each benchmarking round?

(Then again, I've filled up entire disks with Linux perf traces, so maybe I'm biased. :)

Patch-set: 2
Label: Code-Review=0