benchmarks: fix usage of obsolete tracing functions
Currently it's not possible to build the project due to unstable
external API introduced in one of the recent patches.

As for Go 1.5, execution tracing is implemented in a separate
package named "trace". Signature of functions has changed as well.

This patch makes tracing functionality compatible with the stable
Go 1.5 release.

Change-Id: I92959e9cbe76e5ffe4c6cba70587159d797ed1c7