all: report in standard benchmark format
This switches all benchmarks to report in the standard Go benchmark
format [1], making it possible to run and analyze these benchmarks
using tools built for standard Go benchmarks.

As part of this, we lift the repetition loop to the top level to make
it possible to report the results after each run.

[1] https://github.com/golang/proposal/blob/master/design/14313-benchmark-format.md

Change-Id: I542e0f4f6fcfa7783ee37bcdc7edf644201b0b60