Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Bench binary also served as benchmark discovery mechanism for perf dashboard. Any there any plans for any substitution?

No specific plans, though I think the perf dashboard revival has several harder problems to solve first. :)

I can think of a few solutions, though. Honestly, I'd be perfectly okay with hard-coding the list since it doesn't exactly change often. But an easy discovery mechanism would be to find the main packages in the benchmarks repo. Or we could put a text file in the benchmarks repo that lists which directories are benchmarks.

Patch-set: 1