reflect: add a copy of BenchmarkMap
This is a copy of reflect_test.BenchmakMap from the Go standard library,
added in http://golang.org/cl/345486.

This is added to x/benchmarks only temporarily as a sample 'go test'
benchmark to use with benchmarking tools. This benchmark in particular
was selected because it had a significant performance improve with a
recent upstream change (http://golang.org/cl/345486).

Change-Id: I2bff47ffbaab1d471a37ec3cba4bea88cc78d8ed