cmd/bent: refactor common benchmark info into "suites.toml"; add versions
Common fields in benchmark*toml were a headache, especially w/
the addition of versions.  This was a big improvement, all the
common info (build flags, gc env, etc) can appear there.

Also added interpretation of "temporary pattern" for configured
"RunDir" for tests that expect to run in a writeable directory
(and also w/o testdata, which seems to be the case for at least
one bench).

Change-Id: Id80695c11466ec86810107c1a4acefcde991bff9