sweet: add sweet benchmarks
These aren't yet fully functional because the assets aren't yet public.
Later CLs will enable use of assets and add tests. This is basically
just an export of the internal version, with import paths updated so
that it builds.

It also makes a very tiny change to the way the help string is printed
to avoid issues with go vet, and places all the gVisor-related code
behind a Linux build tag so that the subdirectory at least builds on
other platforms.

Change-Id: Ia0d733accfba73fa64dac3b0bd31ee82cad6d92f