sweet: validate that assets are unmodified instead defensively copying
Currently Sweet defensively copies assets for each run of a benchmark,
but this is really wasteful in terms of time. Instead, this CL makes
Sweet build an index of mod times that it then uses to check for
modifications of any of the asset files.

In practice, no benchmark actually modifies asset files. However, some
benchmarks need to open assets as read-write.

Change-Id: I9ccc46c3b55109c6916bc61e3c3fb6c047540546