cmd/bent: move toolchain key print to lower layer
CL 462956 added logging of the toolchain key to properly annotate build
/ after-build benchmarks. This worked fine for the perf dashboard, but
results also go to a file, which this logging did not do. Additionally,
it logged even in non-verbose mode, when benchmark results are not.

Move this logging down near the benchmarks themselves to fix these
inconsistencies.

For golang/go#57770.

Change-Id: I6502092bd93a2dd6932018989c69fba5b51adff8