cmd/bent: add -X flag for benchmark/library-modification experiments
Normally, for no-surprises reproducibility, bent always starts
the build directory from scratch.  This change adds a flag that
turns off that clean-start behavior, so that local edits to
build/*/go.mod for various benchmarks are not reverted.  For
example, if a change to Go breaks a library used by a benchmark,
replace directives in go.mod allow use of a modified library.

Change-Id: Ida8d3776eae081cc4fae63a0489cbb0579b7c5a8