sweet: force -pgo=off for non-PGO builds
As of CL 497455, cmd/compile has a default.pgo that is applied
automatically. This means that `sweet run -pgo -run go-build` is
actually going to compare default.pgo vs a new profile, rather than no
pgo vs a new profile, which is what folks will expect.

Thus when not in a PGO build, force PGO off.

Change-Id: Ica1da2d2d8a4348cc3b8fa2e7b5bdbe9ac00e2ee