sweet: make tile38 and go-build with -short even shorter
-short is not meant to be used for benchmarking; just for testing and
debugging.

Change tile38 to run with a bare (empty) database, since a substantial
amount of time is spent on loading the database from disk. The lack of a
database also means that the tile38 server uses substantially less
memory, so we can let it run in parallel with more tasks in the
end-to-end test.

Change go-build to not do a first build of the target package during
Sweet's build step: it means the results won't be as useful since
they'll include downloading packages, but in -short mode it doesn't
matter.

Change-Id: I9e3559c4077fff696a374b97f91491783a26e5ac