sweet: build cmd/compile, cmd/link with "-a" for go-build benchmark
For the go-build benchmark, make sure to build the tools used for the
runs with "go build -a" and not "go build". This is intended to help
with situations where we're testing a new GOEXPERIMENT that perturbs
compiler behavior.

For example, let's suppose that I write some code in the compiler that
produces better loop alignment (to avoid icache issues) and I enable
the optimization under GOEXPERIMENT=fastloops.

If I then do two "sweet" runs one with a toolchain built with
GOEXPERIMENT=fastloops and one without the go experiment setting, in
order to get an "applies to apples" comparison I need the
"cmd/compile" binary built by the harness to _not_ have
GOEXPERIMENT=fastloops enabled. Because the sweet harness does a clone
of the Go repo, just a regular "go build" will just relink the binary,
won't recompile everything (as needed).

Change-Id: Id47f0b40664a178ac8c11f9657ee042366d7aaf4